+++ lib/src/log.dart
@@ -2,10 +2,10 @@ import 'package:retro_logger/src/enums/log_type.dart';
 
 export 'package:retro_logger/src/enums/log_type.dart';
 
-/// A class representing a log entry with a level, message, timestamp, origin, and type.
+/// A class representing a log entry with a level, message, timestamp, name, and type.
 class Log {
   /// The unique identifier of the log.
-  final String origin;
+  final String name;
 
   /// The level of the log (e.g., 'info', 'error').
   final String level;
@@ -23,7 +23,7 @@ class Log {
   ///
   /// If [timestamp] is not provided, it defaults to the current date and time.
   Log({
-    required this.origin,
+    required this.name,
     required this.level,
     required this.message,
     required this.type,
@@ -34,14 +34,14 @@ class Log {
   ///
   /// If a field is not provided, the current value of that field is used.
   Log copyWith({
-    String? origin,
+    String? name,
     String? level,
     String? message,
     LogType? type,
     DateTime? timestamp,
   }) {
     return Log(
-      origin: origin ?? this.origin,
+      name: name ?? this.name,
       level: level ?? this.level,
       message: message ?? this.message,
       type: type ?? this.type,
@@ -51,11 +51,11 @@ class Log {
 
   /// Creates a new [Log] instance from a JSON object.
   ///
-  /// The JSON object must contain the 'origin', 'level', 'message', and 'type' fields.
+  /// The JSON object must contain the 'name', 'level', 'message', and 'type' fields.
   /// The 'timestamp' field is optional and defaults to the current date and time if not provided.
   factory Log.fromJson(Map<String, dynamic> json) {
     return Log(
-      origin: json['origin'] as String? ?? '',
+      name: json['name'] as String? ?? '',
       level: json['level'] as String? ?? 'unknown',
       message: json['message'] as String? ?? '',
       type: LogType.values.firstWhere(
@@ -70,10 +70,10 @@ class Log {
 
   /// Converts this [Log] instance to a JSON object.
   ///
-  /// The resulting JSON object contains the 'origin', 'level', 'message', 'type', and 'timestamp' fields.
+  /// The resulting JSON object contains the 'name', 'level', 'message', 'type', and 'timestamp' fields.
   Map<String, dynamic> toJson() {
     return {
-      'origin': origin,
+      'name': name,
       'level': level,
       'message': message,
       'type': type.toString(),
@@ -83,7 +83,7 @@ class Log {
 
   @override
   String toString() {
-    return 'Log(origin: $origin, level: $level, message: $message, type: $type, timestamp: $timestamp)';
+    return 'Log(name: $name, level: $level, message: $message, type: $type, timestamp: $timestamp)';
   }
 
   @override
@@ -91,7 +91,7 @@ class Log {
     if (identical(this, other)) return true;
 
     return other is Log &&
-        other.origin == origin &&
+        other.name == name &&
         other.level == level &&
         other.message == message &&
         other.type == type &&
@@ -100,7 +100,7 @@ class Log {
 
   @override
   int get hashCode {
-    return origin.hashCode ^
+    return name.hashCode ^
         level.hashCode ^
         message.hashCode ^
         type.hashCode ^