+++ test/logger_test.dart
@@ -12,7 +12,7 @@ void main() {
 
     test('should add a log', () {
       final log = Log(
-        origin: 'test: should add a log',
+        name: 'test: should add a log',
         level: 'info',
         message: 'This is a test log',
         timestamp: DateTime.now(),
@@ -27,14 +27,14 @@ void main() {
 
     test('should retrieve logs', () {
       final log1 = Log(
-        origin: 'test: should retrieve logs',
+        name: 'test: should retrieve logs',
         level: 'info',
         message: 'First log',
         timestamp: DateTime.now(),
         type: LogType.other,
       );
       final log2 = Log(
-        origin: 'test: should retrieve logs',
+        name: 'test: should retrieve logs',
         level: 'error',
         message: 'Second log',
         timestamp: DateTime.now(),
@@ -51,7 +51,7 @@ void main() {
 
     test('should clear logs', () {
       final log = Log(
-        origin: 'test: should clear logs',
+        name: 'test: should clear logs',
         level: 'info',
         message: 'This is a test log',
         timestamp: DateTime.now(),